+++ R/association.R
@@ -140,8 +140,8 @@ searchDisease <- function(diseaseName) {
 #'
 #' \donttest{
 #' # search disease
-#' searchDisease("Alzheimer disease")
-#' disId <- "Alzheimer disease"
+#' searchDisease("response to antidepressant")
+#' disId <- "response to antidepressant"
 #'
 #' # retrieve associated SNPs
 #' association <- findMirnaSNPs(obj, disId)